+++ turbopack/crates/turbopack-browser/src/ecmascript/evaluate/chunk.rs
@@ -174,6 +174,13 @@ impl EcmascriptBrowserEvaluateChunk {
         let asset_context = turbopack::get_runtime_asset_context(environment);
 
         let runtime_type = *this.chunking_context.runtime_type().await?;
+        // Detect async modules from the whole-app graph in production. In development, the graph
+        // is per-page. To keep the shared `runtime.js` stable, always include the machinery.
+        let has_async_modules = if matches!(runtime_type, RuntimeType::Production) {
+            !this.module_graph.async_module_info().await?.is_empty()
+        } else {
+            true
+        };
         match runtime_type {
             RuntimeType::Production | RuntimeType::Development => {
                 let runtime_code = turbopack_ecmascript_runtime::get_browser_runtime_code(
@@ -187,6 +194,7 @@ impl EcmascriptBrowserEvaluateChunk {
                     source_maps,
                     this.chunking_context.chunk_loading_global(),
                     this.chunking_context.cross_origin(),
+                    has_async_modules,
                 );
                 code.push_code(&*runtime_code.await?);
             }
